refactor my code
