+++ app/(instructor_dashboard)/dashboard/create-new-course/_components/TabSection/BasicInformationTab.jsx
